Stuff I did wrong: I made the profile_for smarter, by adding another argument "user" to the function and making the userid a random value. This is the solution for Challenge 13. In this challenge, we have to create a role=admin after setting up the functions for profile_for, encrypting and decrypting with AES-ECB. The encrypt function is the oracle.
We have to get the correct blocks of ciphertext and concatenate it to create the role=admin
First, I get the ciphertext for "email=(a*7)&uid=10&role=". I make sure that this is 2 complete
blocks. Then I send email=(a*9)+(admin with pkcs7coding). The second block in the ciphertext
has admin with the correct padding value. If we take the first 2 blocks of the first ciphertext,
and the 2nd block of the second ciphertext and send that for decryption. We get back the user
with the correct role.